Separate some of the logic in the request evidence determination class
